Cleaned up SignalGenerator includes and file structure to align with SimpleGpio.
